+++ webdev/lib/src/command/daemon_command.dart
@@ -12,8 +12,8 @@ import 'package:pedantic/pedantic.dart';
 import '../daemon/app_domain.dart';
 import '../daemon/daemon.dart';
 import '../daemon/daemon_domain.dart';
+import '../logging.dart';
 import '../serve/dev_workflow.dart';
-import '../serve/logging.dart';
 import '../serve/utils.dart';
 import 'configuration.dart';
 import 'shared.dart';
@@ -56,15 +56,13 @@ class DaemonCommand extends Command<int> {
       daemon = Daemon(_stdinCommandStream, _stdoutCommandResponse);
       var daemonDomain = DaemonDomain(daemon);
       setLogHandler((level, message, {verbose}) {
-        daemonDomain.sendEvent(
-            'daemon.logMessage', {'level': '$level', 'message': message});
+        daemonDomain.sendEvent('daemon.log', {'log': message});
       });
       daemon.registerDomain(daemonDomain);
       var configuration =
           Configuration(launchInChrome: true, debug: true, autoRun: false);
       var pubspecLock = await readPubspecLock(configuration);
-      var buildOptions =
-          buildRunnerArgs(pubspecLock, configuration, includeOutput: false);
+      var buildOptions = buildRunnerArgs(pubspecLock, configuration);
       var port = await findUnusedPort();
       workflow = await DevWorkflow.start(
         configuration,