+++ webdev/lib/src/daemon/app_domain.dart
@@ -21,7 +21,9 @@ import 'utilites.dart';
 class AppDomain extends Domain {
   String _appId;
   AppDebugServices _appDebugServices;
+
   DebugService get _debugService => _appDebugServices?.debugService;
+
   VmService get _vmService => _appDebugServices?.webdevClient?.client;
   StreamSubscription<BuildResult> _resultSub;
   StreamSubscription<Event> _stdOutSub;
@@ -36,16 +38,13 @@ class AppDomain extends Domain {
         sendEvent('app.progress', {
           'appId': _appId,
           'id': '$_buildProgressEventId',
-          'message': 'Building',
-          'progressId': 'build',
+          'message': 'Building...',
         });
         break;
       case BuildStatus.failed:
         sendEvent('app.progress', {
           'appId': _appId,
           'id': '$_buildProgressEventId',
-          'message': 'Build Failed',
-          'progressId': 'build',
           'finished': true,
         });
         break;
@@ -53,8 +52,6 @@ class AppDomain extends Domain {
         sendEvent('app.progress', {
           'appId': _appId,
           'id': '$_buildProgressEventId',
-          'message': 'Build Succeeded',
-          'progressId': 'build',
           'finished': true,
         });
         break;
@@ -144,8 +141,10 @@ class AppDomain extends Domain {
     if (_appId != appId) throw ArgumentError.value(appId, 'appId', 'Not found');
     var fullRestart = getBoolArg(args, 'fullRestart') ?? false;
     if (!fullRestart) {
-      throw ArgumentError.value(
-          fullRestart, 'fullRestart', 'We do not support hot reload yet.');
+      return {
+        'code': 1,
+        'message': 'hot reload not yet supported by package:flutter_web',
+      };
     }
     // TODO(grouma) - Support pauseAfterRestart.
     // var pauseAfterRestart = getBoolArg(args, 'pause') ?? false;