+++ webdev/lib/src/serve/handlers/dev_handler.dart
@@ -14,11 +14,11 @@ import 'package:shelf/shelf.dart';
 import 'package:sse/server/sse_handler.dart';
 import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
 
-import '../../serve/chrome.dart';
-import '../../serve/data/run_request.dart';
-import '../../serve/logging.dart';
+import '../../logging.dart';
+import '../chrome.dart';
 import '../data/connect_request.dart';
 import '../data/devtools_request.dart';
+import '../data/run_request.dart';
 import '../data/serializers.dart' as webdev;
 import '../debugger/app_debug_services.dart';
 import '../debugger/devtools.dart';
@@ -108,8 +108,20 @@ class DevHandler {
           return;
         }
 
-        var appServices =
-            await loadAppServices(message.appId, message.instanceId);
+        AppDebugServices appServices;
+        try {
+          appServices =
+              await loadAppServices(message.appId, message.instanceId);
+        } catch (_) {
+          connection.sink.add(
+              jsonEncode(webdev.serializers.serialize(DevToolsResponse((b) => b
+                ..success = false
+                ..error = 'Webdev was unable to connect debug services to your '
+                    'application. Most likely this means you are trying to '
+                    'load in a different Chrome window than was launched by '
+                    'webdev.'))));
+          return;
+        }
 
         // Check if we are already running debug services for a different
         // instance of this app.